+++ apps/neoscan/priv/repo/migrations/20180607000006_address_histories.exs
@@ -6,7 +6,7 @@ defmodule Neoscan.Repo.Migrations.AddressHistories do
       add(:address_hash, :binary, null: false)
       add(:transaction_hash, :binary, null: false)
       add(:asset_hash, :binary, null: false)
-      add(:value, :float, null: false)
+      add(:value, :decimal, null: false)
       add(:block_time,  :naive_datetime, null: false)
 
       timestamps()
